+++ sdk/storage/Azure.Storage.Blobs.Batch/src/BlobBatchClient.cs
@@ -399,7 +399,7 @@ namespace Azure.Storage.Blobs.Specialized
         /// that the operation should be cancelled.
         /// </param>
         /// <returns>A Task representing the update operation.</returns>
-        private static async Task UpdateOperationResponses(
+        private async Task UpdateOperationResponses(
             IList<HttpMessage> messages,
             Response rawResponse,
             Stream responseContent,
@@ -428,7 +428,7 @@ namespace Azure.Storage.Blobs.Specialized
                     if (responses.Length == 1 && responses[0].Status == 400)
                     {
                         // We'll re-process this response as a batch result
-                        BatchRestClient.Service.SubmitBatchAsync_CreateResponse(responses[0]);
+                        BatchRestClient.Service.SubmitBatchAsync_CreateResponse(ClientDiagnostics, responses[0]);
                     }
                     else
                     {
@@ -439,7 +439,7 @@ namespace Azure.Storage.Blobs.Specialized
             catch (InvalidOperationException ex)
             {
                 // Wrap any parsing errors in a RequestFailedException
-                throw BatchErrors.InvalidResponse(rawResponse, ex);
+                throw BatchErrors.InvalidResponse(ClientDiagnostics, rawResponse, ex);
             }
 
             // Update the delayed responses
