+++ sdk/storage/Azure.Storage.Blobs.Batch/src/Models/StorageError.cs
@@ -46,6 +46,6 @@ namespace Azure.Storage.Blobs.Models
         /// <param name="response">The failed response.</param>
         /// <returns>A RequestFailedException.</returns>
         public Exception CreateException(ClientDiagnostics clientDiagnostics, Azure.Response response)
-            => clientDiagnostics.CreateRequestFailedExceptionWithContent(response, TODO, message: Message, content: null, errorCode: Code, additionalInfo: AdditionalInformation);
+            => clientDiagnostics.CreateRequestFailedExceptionWithContent(response, message: Message, content: null, errorCode: Code, additionalInfo: AdditionalInformation);
     }
 }
